scripts/API_package/requirements.txt
requests
pandas
matplotlib.pyplot
numpy
seaborn
datetime
dateutil.relativedelta
